create test suit for testing request-response->request scheme (tracking_wallet)
chore:
* cleanup code
* sort dependencies
* move tests from indexer_internal
* add entrypoint for handling track_wallet
