+++ btc_indexer/crates/config_parser/Cargo.toml
@@ -4,6 +4,7 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+bitcoincore-rpc = { workspace = true }
 config = { workspace = true }
 dotenv = { workspace = true }
 serde = { workspace = true }
@@ -11,4 +12,3 @@ serde_json = { workspace = true }
 strum = { workspace = true }
 thiserror = { workspace = true }
 tracing = { workspace = true }
-bitcoincore-rpc = { workspace = true }
