+++ btc_indexer/crates/indexer_internals/tests/utils/mod.rs
@@ -1,0 +1,2 @@
+pub mod common;
+pub mod mock;
