+++ btc_indexer/crates/server/src/error.rs
@@ -17,7 +17,6 @@ pub enum ServerError {
     TaskCancelled(String),
 }
 
-pub use response_conversion::*;
 mod response_conversion {
     use axum::http::StatusCode;
