+++ btc_indexer/infrastructure/configuration/local.toml
@@ -1,4 +1,4 @@
 [application]
 
 [btc_indexer]
-update_interval_millis=500
+update_interval_millis = 500
