+++ btc_indexer/infrastructure/configuration/production.toml
@@ -3,4 +3,4 @@ http_server_ip = "0.0.0.0"
 http_server_port = 8080
 
 [btc_indexer]
-update_interval_millis=500
+update_interval_millis = 500
