+++ crates/global_utils/Cargo.toml
@@ -4,6 +4,6 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
-tracing = {workspace = true }
-tracing-appender = {workspace = true }
-tracing-subscriber = {workspace = true }
+tracing = { workspace = true }
+tracing-appender = { workspace = true }
+tracing-subscriber = { workspace = true }
